+++ opts/throttledevice.go
