+++ opts/weightdevice.go
